some additions

+++ src/HW_5_1.java
@@ -60,7 +60,7 @@ public class HW_5_1 {
 
         int[] array4 = {1, 2, 3, 4, 5, 6, 7, 8, 9};
 
-        int sumarray4 = 0;
+        double sumarray4 = 0;
 
         for (int i = 0; i < array4.length; i++) {
             sumarray4 += array4[i];
